Fix message length parsing on 32-bit platforms
On 32-bit platforms, int(uint32) for values above MaxInt32 silently
wraps negative, bypassing the < 4 length check and producing a
negative bodyLen. Cast through int32 first so that large uint32
values become negative and are properly rejected.

+++ pgproto3/backend.go
@@ -123,7 +123,7 @@ func (b *Backend) ReceiveStartupMessage() (FrontendMessage, error) {
 	if err != nil {
 		return nil, err
 	}
-	msgSize := int(binary.BigEndian.Uint32(buf) - 4)
+	msgSize := int(int32(binary.BigEndian.Uint32(buf)) - 4)
 
 	if msgSize < minStartupPacketLen || msgSize > maxStartupPacketLen {
 		return nil, fmt.Errorf("invalid length of startup packet: %d", msgSize)
@@ -176,7 +176,7 @@ func (b *Backend) Receive() (FrontendMessage, error) {
 
 		b.msgType = header[0]
 
-		msgLength := int(binary.BigEndian.Uint32(header[1:]))
+		msgLength := int(int32(binary.BigEndian.Uint32(header[1:])))
 		if msgLength < 4 {
 			return nil, fmt.Errorf("invalid message length: %d", msgLength)
 		}
